moved sources-summary page into sources-detail view directory
