+++ client/src/app/views/sources/source-add/source-add.page.less
@@ -1,0 +1,1 @@
+@import 'themes/overrides/entity-page-header.overrides.less';
