+++ client/src/app/views/sources/sources-detail/sources-summary/sources-summary.page.html
